Phase 4: Extract API error/interceptor classes, remove dead storage code
- Extract ApiError, RateLimitError, normalizeApiError into config/api/errors.js
- Extract request/response interceptors into config/api/interceptors.js
- api.js reduced from 406 to 293 lines (thin orchestration)
- Remove duplicate NotificationBell re-exports (common/, Layout/)
- Remove unused store files (useAuthStore, useNotificationStore, useMyEventsStore)
