--- src/store/useMyEventsStore.js
@@ -1,9 +1,0 @@
-import { create } from 'zustand';
-
-export const useMyEventsStore = create((set) => ({
-  myEvents: [],
-  setMyEvents: (events) => set({ myEvents: events }),
-  addRegistration: (event, formData, registrationId, qrToken) => set((state) => ({
-    myEvents: [...state.myEvents, { ...event, registrationData: formData, registrationId, qrToken }]
-  })),
-}));
